+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
  - return 404 http status code when resource not exists 
  - add auto-tag script
+ - responses accept links
+ - add self link in response
+ - update auto-tag-script
 
 # Version 1.1